Fix vogar's idea of quoting.
This patch switches us over to using "sh -c" on the host and
"adb shell sh -c" on the device for consistency. Before we were using
exec(2) directly on the host and relying on a bug in adb shell on the
device. This patch stops exploiting that bug (which was fixed as bug
http://b/20323053) and stops using exec(2) locally to keep us honest.

We should probably remove SshTarget, but that's a job for another patch.

Bug: http://b/19734868
Bug: http://b/20323053
Change-Id: I9a74050e10502ff062fe404f8359aecdb82afbe2